feat: Score컴포넌트 분리
